requirements.txt
# pip install -r requirements.txt

pyzmq
gevent